Build GStreamer and plugins ourselves
We need a couple of plugins that aren't available in the GStreamer
version supplied by the runtime, so build the latest stable GStreamer
and the necessary plugins, so we have all the bug fixes.

This fixes the runtime update which left out (at least) a plugin to
mux AAC/M4A files.

Fixes: 8832104ea78be50e8bf9fda798facdc15432a660